Consolidate the cqf-cqlType extension reader into interval-utils
Three copies of the extension URL and reader had accumulated: the
datetime interval extractor (from #109), the numeric interval
extractor, and the part-form point-type derivation in ResultExtractor.
A single declaredCqlType helper in interval-utils now returns the raw
declared type string; interpreting it stays with each caller, since
they accept different type families. No behavior change.
